vehicra-backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic[email]
python-multipart
boto3